+++ sshd-core/src/test/java/org/apache/sshd/client/ClientTest.java
@@ -208,7 +208,7 @@ public class ClientTest extends BaseTestSupport {
                         return new ChannelSession() {
                             @SuppressWarnings("synthetic-access")
                             @Override
-                            public OpenFuture open(int recipient, long rwsize, long rmpsize, Buffer buffer) {
+                            public OpenFuture open(long recipient, long rwsize, long rmpsize, Buffer buffer) {
                                 try {
                                     channelLatch.await();
                                 } catch (InterruptedException e) {
@@ -342,7 +342,7 @@ public class ClientTest extends BaseTestSupport {
 
     @Test
     public void testClientStillActiveIfListenerExceptions() throws Exception {
-        Map<String, Integer> eventsMap = new TreeMap<>(String.CASE_INSENSITIVE_ORDER);
+        Map<String, Long> eventsMap = new TreeMap<>(String.CASE_INSENSITIVE_ORDER);
         Collection<String> failuresSet = new TreeSet<>(String.CASE_INSENSITIVE_ORDER);
         Logger log = LoggerFactory.getLogger(getClass());
         client.addChannelListener(new ChannelListener() {
@@ -377,7 +377,7 @@ public class ClientTest extends BaseTestSupport {
             }
 
             private void handleChannelEvent(String name, Channel channel) {
-                int id = channel.getId();
+                long id = channel.getId();
                 synchronized (eventsMap) {
                     if (eventsMap.put(name, id) != null) {
                         return; // already generated an exception for this event
@@ -1475,9 +1475,9 @@ public class ClientTest extends BaseTestSupport {
             channels.add(session.createChannel(Channel.CHANNEL_EXEC, getCurrentTestName()));
             channels.add(session.createChannel(Channel.CHANNEL_SHELL, getClass().getSimpleName()));
 
-            Set<Integer> ids = new HashSet<>(channels.size());
+            Set<Long> ids = new HashSet<>(channels.size());
             for (ClientChannel c : channels) {
-                int id = c.getId();
+                long id = c.getId();
                 assertTrue("Channel ID repeated: " + id, ids.add(id));
             }
         } finally {
